Separating auth verification form
